fix classes

+++ src/main/java/com/app/back/banco/tiposCuentasHerencia/CuentaPesoCL.java
@@ -6,7 +6,7 @@ import com.app.back.banco.Cuenta;
 public class CuentaPesoCL extends Cuenta { //Aplicamos herencia con el extends y decimos cual es el padre
     private static final double TASA_CAMBIO_USD_A_CLP = 850.0; //Es una constante que podremos usar siempre con esta clase
 
-    public CuentaPesoCL(int numeroCuenta, long idCliente, double saldoActual) {
+    public CuentaPesoCL(int numeroCuenta, long idCliente, int saldoActual) {
         super(numeroCuenta, idCliente, saldoActual);
     }
